Introduce leak safe API
Previously rakaly_eu4_melt would modify a the given out pointer to point
to rust allocated chunk of data. There is no way for callers to
deallocate this data safely, so right now causing memory leaks.

This can be seen in the sample.c. When ran under valgrind, it will
complain about memory leaks:

```
LEAK SUMMARY:
   definitely lost: 5,056,975 bytes in 1 blocks
   indirectly lost: 0 bytes in 0 blocks
     possibly lost: 75,353,864 bytes in 1 blocks
   still reachable: 0 bytes in 0 blocks
        suppressed: 0 bytes in 0 blocks
```

The fix is to introduce an API that that allows the caller to free the
allocated structure (`MeltedBuffer`).

In addition the fix is to also provide a function that copies the melted
data to a caller provided buffer that has the correct length. An
alternative design was to provide a pointer to the melted buffer data
directly but then we'd run the risk that the caller could hold onto the
data after `rakaly_melt_free` was called.

After updating sample.c to use the new API, valgrind memcheck passed.