Avoid forcing a null terminator to be written to the provided buffer
There's an issue where `rakaly_melt_write_data` writes a null terminator
into the provided buffer even though the caller may not want a null
terminator.

This is apparent in the following c++ wrapper:

```cpp
size_t len = rakaly_melt_data_length(melt) + 1;
std::string str(len, ' ');
size_t written = rakaly_melt_write_data(melt, str.data(), len);
```

In the above the string had to explicitly make room for the null
terminator and since the null terminator is now considered a part of the
string, one must now explicitly trim it.

The fix is to make it the caller's responsibility for ensuring the
buffer they are working with is null terminated. This in turn, makes the
code for everyone more succint and less error prone.